Backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
mutagen==1.46.0
celery==5.3.1
redis==4.6.0
stripe==6.8.0
python-jose[cryptography]==3.3.0
cryptography==41.0.4
bcrypt==4.0.1
email-validator==2.1.0
marshmallow==3.20.1
apispec==6.3.0
flask-marshmallow==1.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
SQLAlchemy-Utils==0.41.1
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Mail==0.9.1
python-dateutil==2.8.2
schedule==1.2.0
APScheduler==3.10.4
